Move ChatMessadapter into another package
